second increment, third increment, fourth increment.

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\DB;
 class UserController extends Controller
 {
     public function index()
@@ -18,7 +19,7 @@ class UserController extends Controller
     {
         //validate the fields
         $request->validate([
-            'name' => 'required|max:255',
+            'name' => 'required|string|max:255',
             'email' => 'required|unique:users|email|max:255',
             'password' => 'required|between:8,255|confirmed',
             'password_confirmation' => 'required'
@@ -38,7 +39,7 @@ class UserController extends Controller
     {
         //validate the fields
         $request->validate([
-            'name' => 'required|max:255',
+            'name' => 'required|max:255|string',
             'email' => 'required|email|max:255',
             'password' => 'confirmed'
         ]);
@@ -52,7 +53,17 @@ class UserController extends Controller
     }
     public function destroy(User $user)
     {
-        $user->delete();
-        return redirect('/users');
+        $id = $user->id;
+        $users = DB::select('select u.id
+                            from receipts r, users u
+                            where u.id ='.$id.'
+                            and u.id = r.userId
+                            ');
+        if($users==null){
+            $user->delete();
+            return redirect('/users');
+        }else{
+            return redirect('/users')->with('Message','Error when deleting the user, because it is linked to a receipt.');
+        }
     }
 }
